+++ Changelog.md
@@ -6,6 +6,7 @@ Released on ?.
 
 ✨ New Features / Improvements
 
+- Added the "Terrain slope direction" map overlay
 - Using CesiumJS version 1.131.0
 
 🐛 Bugfixes